move download_tgz, fetch_from_git to untils
